Three section approach success
The rightmost section is where you type telomare code, the middle
section is the parsed version of that telomare code with checkboxes in
each node to be selected and evaluated to an IExpr (evaluation which shows on the
leftmost section).

Still need to make checkboxes uncheck when a new one is checked.